file reorg, put support.py to ops_metal
